setup policy

+++ app/Http/Controllers/Product/CartController.php
@@ -16,28 +16,34 @@ class CartController extends Controller
         $data = [];
         $grandTotal=0;
         foreach(Auth()->user()->cart as $cart){
+            if($cart->product){//IF PRODUCT EXIST
+                $fullPrice = $cart->product->price;      // price that not discount
+                $dis = $cart->product->discount;         // percent of discount
+                $qua = $cart->quantity;                  // quantity of product that in same category
+                $total = $fullPrice*$qua*(100-$dis)/100; // total price for each category
 
-            $fullPrice = $cart->product->price;      // price that not discount
-            $dis = $cart->product->discount;         // percent of discount
-            $qua = $cart->quantity;                  // quantity of product that in same category
-            $total = $fullPrice*$qua*(100-$dis)/100; // total price for each category
+                $data[] = [
+                    'id'                 => $cart->id,
+                    'product_id'         => $cart->product->id,
+                    'price'              => $fullPrice,
+                    'discount(%)'        => $dis,
+                    'product_image_path' => $cart->product->product_image_path,
+                    'quantity'           => $qua,
+                    'Total($)'           => $total,
+                    'increase (get method)'       => PHP_URL_HOST .route('cart.show',$cart->id),
+                    'decrease (PUT|PATCH method)' => PHP_URL_HOST .route('cart.update',$cart->id),
+                    'delete (delete method)'      => PHP_URL_HOST .route('cart.destroy',$cart->id)
+                ];
 
-            $data[] = [
-                'id'                 => $cart->id,
-                'product_id'         => $cart->product->id,
-                'price'              => $fullPrice,
-                'discount(%)'        => $dis,
-                'product_image_path' => $cart->product->product_image_path,
-                'quantity'           => $qua,
-                'Total($)'           => $total,
-            ];
-
-            $grandTotal += $total;
+                $grandTotal += $total;
+            }
+            else  $cart->delete();
+            
         }
         
         return response([
             'Product in cart' => $data,
-            'Grand Total($)'  => $grandTotal  
+            'Grand Total($)'  => '$grandTotal',   
         ]);
     }
 
@@ -61,7 +67,7 @@ class CartController extends Controller
             ]);
         }
 
-        //add item t cart
+        //add item to cart
         Cart::create([
             'user_id'   => Auth()->user()->id,
             'product_id'=> $request->product_id,
@@ -74,8 +80,17 @@ class CartController extends Controller
 
     }
 
+    //increase
+    public function show(Cart $cart){
+        $this->authorize('update',$cart);//policy
+        $cart->quantity += 1;
+        $cart->save();
+    }
+
+    //decrease item
     public function update(Cart $cart){
-        // $request->validated();
+        
+        $this->authorize('update',$cart);//policy
         
         if($cart->quantity===1){
             $cart->delete();
@@ -95,6 +110,8 @@ class CartController extends Controller
 
     //deleted item from cart
     public function destroy(Cart $cart){
+
+        $this->authorize('update',$cart);//policy
         $cart->delete();
         return [
             'message' => 'item deleted'
